[REFACTOR] 🔧 use cases
