+++ app/use_cases/api/v1/system/memory_storage/application_storage_use_case.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Api
+  module V1
+    module System
+      module MemoryStorage
+        class ApplicationStorageUseCase < ApplicationUseCase
+          KEY_BASE = 'APP::Repository::'
+          TTL = 1.hour
+
+          private_constant :KEY_BASE
+          private_constant :TTL
+        end
+      end
+    end
+  end
+end
